WIP - remove `id` parameter from `checkDocType`
First off these changes were not tested so they might not work. As the
name of the branch states: this is on-going development.

In order to make a working "sharing" demo we had to hack the permissions
checking system and a parameter `id` was added. Since, this parameter was
removed in a previous commit and this change was not propagated
throughout all of the code. This commit aims at correcting that.